# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Personal Digital Twin Assistant API",
    "description": "Backend API for the Personal Digital Twin productivity assistant",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/graph": {
      "get": {
        "tags": [
          "Knowledge Graph"
        ],
        "summary": "Get Graph Data",
        "description": "Get all items as nodes and all item_links as edges for the knowledge graph.",
        "operationId": "get_graph_data_api_graph_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphData"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/graph/links": {
      "post": {
        "tags": [
          "Knowledge Graph"
        ],
        "summary": "Create Link",
        "description": "Create a manual link between two items.",
        "operationId": "create_link_api_graph_links_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemLinkCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemLinkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/graph/analyze": {
      "post": {
        "tags": [
          "Knowledge Graph"
        ],
        "summary": "Analyze Graph",
        "description": "Analyze the user's items and create AI-suggested graph links.",
        "operationId": "analyze_graph_api_graph_analyze_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphAnalyzeResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/graph/links/{link_id}": {
      "delete": {
        "tags": [
          "Knowledge Graph"
        ],
        "summary": "Delete Link",
        "description": "Delete a link between two items.",
        "operationId": "delete_link_api_graph_links__link_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "link_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Link Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GraphAnalyzeResponse": {
        "properties": {
          "suggested_links": {
            "items": {
              "$ref": "#/components/schemas/ItemLinkResponse"
            },
            "type": "array",
            "title": "Suggested Links"
          },
          "created_count": {
            "type": "integer",
            "title": "Created Count",
            "default": 0
          },
          "skipped_count": {
            "type": "integer",
            "title": "Skipped Count",
            "default": 0
          }
        },
        "type": "object",
        "title": "GraphAnalyzeResponse",
        "description": "Response returned after graph analysis creates or reuses links."
      },
      "GraphData": {
        "properties": {
          "nodes": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Nodes"
          },
          "links": {
            "items": {
              "$ref": "#/components/schemas/ItemLinkResponse"
            },
            "type": "array",
            "title": "Links"
          }
        },
        "type": "object",
        "required": [
          "nodes",
          "links"
        ],
        "title": "GraphData",
        "description": "Schema for the knowledge graph data"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ItemLinkCreate": {
        "properties": {
          "source_id": {
            "type": "integer",
            "title": "Source Id"
          },
          "target_id": {
            "type": "integer",
            "title": "Target Id"
          },
          "link_type": {
            "type": "string",
            "enum": [
              "subtask_of",
              "relates_to",
              "blocks",
              "updates"
            ],
            "title": "Link Type",
            "default": "relates_to"
          }
        },
        "type": "object",
        "required": [
          "source_id",
          "target_id"
        ],
        "title": "ItemLinkCreate",
        "description": "Schema for creating a link between two items"
      },
      "ItemLinkResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "source_id": {
            "type": "integer",
            "title": "Source Id"
          },
          "target_id": {
            "type": "integer",
            "title": "Target Id"
          },
          "link_type": {
            "type": "string",
            "enum": [
              "subtask_of",
              "relates_to",
              "blocks",
              "updates"
            ],
            "title": "Link Type"
          },
          "weight": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight"
          },
          "ai_reasoning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Reasoning"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source_id",
          "target_id",
          "link_type"
        ],
        "title": "ItemLinkResponse",
        "description": "Schema for item link response"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}